T79830141 Unbreak haddock
It looks like our customhook is not being called and haddock 8.8 is producing
empty docs. When --hoogle is given, the empty set of docs triggers an
error as seen in T79564095.
This is due to the redundant depanal call.

- Without the call, the hook gets called and everything works, except for
a test failure related to Pattern synonyms in Hoogle
there are several failures across all the test suites - around 10% of
the tests fail.
- With the call, the hook does not get called. I can see that the modules
are being recompiled by using load' and a custom logger, but our hook is
not used

The reason for this behaviour is stated in the depanal Haddocks:

> Note that each ModSummary in the module graph caches its DynFlags. These
DynFlags are determined by the current session DynFlags and theOPTIONS
and LANGUAGE pragmas of the parsed module. Thus if you want changes to
the DynFlags to take effect you need to call this function again.

Since Hooks live in the DynFlags, calling depanal with the empty hooks
captures them, and our custom hooks get ignored.

The solution is to remove the explicit call to depanal, since it's not
needed as load calls it anyway.

Note that the testsuite is buggy and succeeds with the depanal call, but
presents some failures without it. This is because in the former case,
haddock is not producing any output artifacts, and the test runner
succeeds vacuously (because it doesn't check that the expected artifacts
are produced, it only compares the existing artifacts with the fixture
ones).